test: extract rollout suites
