+++ crates/agent-engine/src/rollout.rs
@@ -1196,1071 +1196,5 @@ impl Clone for RolloutRecorder {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::io;
-    use std::path::Path;
-    use std::pin::Pin;
-    use std::task::{Context, Poll};
-    use tempfile::TempDir;
-    use tokio::io::AsyncWrite;
-    use tokio::time::{Duration, Instant, sleep};
-
-    #[derive(Default)]
-    struct FlushFailWriter {
-        buffer: Vec<u8>,
-    }
-
-    impl AsyncWrite for FlushFailWriter {
-        fn poll_write(
-            mut self: Pin<&mut Self>,
-            _cx: &mut Context<'_>,
-            buf: &[u8],
-        ) -> Poll<io::Result<usize>> {
-            self.buffer.extend_from_slice(buf);
-            Poll::Ready(Ok(buf.len()))
-        }
-
-        fn poll_flush(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-            Poll::Ready(Err(io::Error::other("synthetic flush failure")))
-        }
-
-        fn poll_shutdown(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-            Poll::Ready(Ok(()))
-        }
-    }
-
-    #[test]
-    fn test_rollout_recorder_creation() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new().unwrap();
-            let recorder =
-                RolloutRecorder::new_in_dir("/proc/123", "gemini-2.0-flash", temp_dir.path()).await;
-            assert!(recorder.is_ok());
-
-            let recorder = recorder.unwrap();
-            let path = recorder.path();
-            assert!(
-                path.file_name()
-                    .unwrap()
-                    .to_string_lossy()
-                    .starts_with("rollout-")
-            );
-            assert!(path.to_string_lossy().ends_with(".jsonl"));
-            let items = RolloutRecorder::load_history(path).await.unwrap();
-            assert!(matches!(
-                items.first(),
-                Some(RolloutItem::AgentMachineMeta(meta)) if meta.process_path == "/proc/123"
-            ));
-
-            // Clean up - remove the created file
-            let _ = fs::remove_file(path).await;
-        });
-    }
-
-    #[test]
-    fn test_rollout_recorder_creation_records_explicit_cwd() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new().unwrap();
-            let explicit_cwd = Path::new("/mnt/source/src");
-
-            let recorder = RolloutRecorder::new_in_dir_with_cwd(
-                "test-machine-cwd",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-                Some(explicit_cwd),
-            )
-            .await
-            .unwrap();
-
-            let items = RolloutRecorder::load_history(recorder.path())
-                .await
-                .unwrap();
-            match &items[0] {
-                RolloutItem::AgentMachineMeta(meta) => {
-                    assert_eq!(meta.cwd, "/mnt/source/src");
-                }
-                _ => panic!("Expected AgentMachineMeta"),
-            }
-
-            let _ = fs::remove_file(recorder.path()).await;
-        });
-    }
-
-    #[test]
-    fn test_record_message_flushes() {
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        rt.block_on(async {
-            let temp_dir = TempDir::new().unwrap();
-            let recorder = RolloutRecorder::new_in_dir(
-                "test-machine-flush",
-                "gemini-2.0-flash",
-                temp_dir.path(),
-            )
-            .await
-            .unwrap();
-            recorder
-                .record_message("user", Some("Hello"), None)
-                .await
-                .unwrap();
-
-            let start = Instant::now();
-            let mut found = false;
-            while start.elapsed() < Duration::from_secs(1) {
-                if let Ok(content) = fs::read_to_string(recorder.path()).await
-                    && content.contains("\"type\":\"message\"")
-                    && content.contains("Hello")
-                {
-                    found = true;
-                    break;
-                }
-                sleep(Duration::from_millis(10)).await;
-            }
-
-            assert!(found, "Expected message to be flushed to rollout file");
-        });
-    }
-
-    #[tokio::test]
-    async fn test_record_tape_message_persists_rich_message() {
-        let temp_dir = TempDir::new().unwrap();
-        let recorder =
-            RolloutRecorder::new_in_dir("test-rich-message", "gemini-2.0-flash", temp_dir.path())
-                .await
-                .unwrap();
-
-        let message = crate::tape::Message::Assistant {
-            parts: vec![
-                crate::tape::ContentPart::thinking("internal reasoning"),
-                crate::tape::ContentPart::text("final answer"),
-            ],
-            tool_requests: vec![],
-        };
-        recorder.record_tape_message(&message).await.unwrap();
-
-        let items = RolloutRecorder::load_history(recorder.path())
-            .await
-            .unwrap();
-        let restored = items.into_iter().find_map(|item| match item {
-            RolloutItem::Message(msg) => msg.message,
-            _ => None,
-        });
-
-        let restored = restored.expect("expected rich message payload");
-        assert_eq!(restored.non_thinking_text_content(), "final answer");
-        assert_eq!(
-            restored.thinking_content().as_deref(),
-            Some("internal reasoning")
-        );
-    }
-
-    #[tokio::test]
-    async fn test_record_checkpoint_persists_knowledge_root() {
-        let temp_dir = TempDir::new().unwrap();
-        let recorder = RolloutRecorder::new_in_dir(
-            "test-root-checkpoint",
-            "gemini-2.0-flash",
-            temp_dir.path(),
-        )
-        .await
-        .unwrap();
-
-        recorder
-            .record_checkpoint_with_knowledge_root(
-                "turn-1",
-                "turn_completed",
-                "turn completed",
-                None,
-                "sha256:abc123",
-            )
-            .await
-            .unwrap();
-
-        let items = RolloutRecorder::load_history(recorder.path())
-            .await
-            .unwrap();
-        let checkpoint = items.into_iter().find_map(|item| match item {
-            RolloutItem::Checkpoint(checkpoint) => Some(checkpoint),
-            _ => None,
-        });
-
-        let checkpoint = checkpoint.expect("checkpoint should be persisted");
-        assert_eq!(checkpoint.checkpoint_id, "turn-1");
-        assert_eq!(checkpoint.knowledge_root.as_deref(), Some("sha256:abc123"));
-    }
-
-    #[tokio::test]
-    async fn test_load_history() {
-        let temp_dir = TempDir::new().unwrap();
-        let recorder = RolloutRecorder::new_in_dir("/proc/7", "gemini-2.0-flash", temp_dir.path())
-            .await
-            .unwrap();
-        recorder
-            .record_tape_message(&crate::tape::Message::user("Hello"))
-            .await
-            .unwrap();
-        recorder
-            .record_tool_call(
-                "test_tool",
-                serde_json::json!({}),
-                serde_json::json!({}),
-                true,
-            )
-            .await
-            .unwrap();
-
-        let items = RolloutRecorder::load_history(recorder.path())
-            .await
-            .unwrap();
-        assert_eq!(items.len(), 3);
-
-        // Verify first item is machine meta
-        match &items[0] {
-            RolloutItem::AgentMachineMeta(meta) => {
-                assert_eq!(meta.process_path, "/proc/7");
-                assert!(!meta.rollout_id.is_empty());
-                assert_eq!(meta.model, "gemini-2.0-flash");
-            }
-            _ => panic!("Expected AgentMachineMeta"),
-        }
-
-        // Verify second item is message
-        match &items[1] {
-            RolloutItem::Message(msg) => {
-                assert_eq!(msg.role, "user");
-                assert_eq!(msg.content, Some("Hello".to_string()));
-                assert!(msg.tool_name.is_none());
-            }
-            _ => panic!("Expected Message"),
-        }
-
-        // Verify third item is tool call
-        match &items[2] {
-            RolloutItem::ToolCall(tool) => {
-                assert_eq!(tool.name, "test_tool");
-                assert!(tool.success);
-            }
-            _ => panic!("Expected ToolCall"),
-        }
-    }
-
-    #[tokio::test]
-    async fn test_persist_batch_writes_compaction_attempt_and_summary_together() {
-        let temp_dir = TempDir::new().unwrap();
-        let recorder = RolloutRecorder::new_in_dir(
-            "test-compaction-batch",
-            "gemini-2.0-flash",
-            temp_dir.path(),
-        )
-        .await
-        .unwrap();
-        let attempt = CompactionAttemptSnapshot {
-            attempt_id: "attempt-123".to_string(),
-            submission_id: Some("sub-456".to_string()),
-            request: alan_agent_protocol::CompactionRequestMetadata {
-                mode: alan_agent_protocol::CompactionMode::Manual,
-                trigger: CompactionTrigger::Manual,
-                reason: CompactionReason::ExplicitRequest,
-                focus: Some("preserve todos".to_string()),
-            },
-            result: CompactionResult::Retry,
-            pressure_level: None,
-            memory_flush_attempt_id: None,
-            input_messages: Some(12),
-            output_messages: Some(4),
-            input_prompt_tokens: Some(900),
-            output_prompt_tokens: Some(300),
-            retry_count: 1,
-            tape_mutated: true,
-            warning_message: None,
-            error_message: None,
-            failure_streak: None,
-            reference_context_revision_before: Some(3),
-            reference_context_revision_after: Some(3),
-            timestamp: "2026-01-29T14:31:00Z".to_string(),
-        };
-        let compacted = CompactedItem {
-            message: "Summary after retry".to_string(),
-            attempt_id: Some(attempt.attempt_id.clone()),
-            trigger: Some(CompactionTrigger::Manual),
-            reason: Some(CompactionReason::ExplicitRequest),
-            focus: Some("preserve todos".to_string()),
-            input_messages: Some(12),
-            output_messages: Some(4),
-            input_tokens: Some(900),
-            output_tokens: Some(300),
-            duration_ms: Some(42),
-            retry_count: Some(1),
-            result: Some(CompactionResult::Retry),
-            reference_context_revision: Some(3),
-            timestamp: "2026-01-29T14:31:01Z".to_string(),
-        };
-
-        recorder
-            .persist_batch(vec![
-                RolloutItem::CompactionAttempt(attempt.clone()),
-                RolloutItem::Compacted(compacted.clone()),
-            ])
-            .await
-            .unwrap();
-
-        let items = RolloutRecorder::load_history(recorder.path())
-            .await
-            .unwrap();
-        let persisted_attempt = items.iter().find_map(|item| match item {
-            RolloutItem::CompactionAttempt(attempt) => Some(attempt),
-            _ => None,
-        });
-        let persisted_compacted = items.iter().find_map(|item| match item {
-            RolloutItem::Compacted(compacted) => Some(compacted),
-            _ => None,
-        });
-
-        assert_eq!(persisted_attempt, Some(&attempt));
-        assert_eq!(
-            persisted_compacted.map(|item| item.attempt_id.as_deref()),
-            Some(Some("attempt-123"))
-        );
-        assert_eq!(
-            persisted_compacted.map(|item| item.message.as_str()),
-            Some("Summary after retry")
-        );
-    }
-
-    #[tokio::test]
-    async fn test_persist_items_and_flush_propagates_flush_error() {
-        let attempt = CompactionAttemptSnapshot {
-            attempt_id: "attempt-flush-failure".to_string(),
-            submission_id: None,
-            request: alan_agent_protocol::CompactionRequestMetadata {
-                mode: alan_agent_protocol::CompactionMode::Manual,
-                trigger: CompactionTrigger::Manual,
-                reason: CompactionReason::ExplicitRequest,
-                focus: None,
-            },
-            result: CompactionResult::Failure,
-            pressure_level: None,
-            memory_flush_attempt_id: None,
-            input_messages: Some(4),
-            output_messages: None,
-            input_prompt_tokens: Some(256),
-            output_prompt_tokens: None,
-            retry_count: 0,
-            tape_mutated: false,
-            warning_message: None,
-            error_message: Some("synthetic".to_string()),
-            failure_streak: Some(1),
-            reference_context_revision_before: Some(2),
-            reference_context_revision_after: None,
-            timestamp: "2026-03-18T00:00:00Z".to_string(),
-        };
-        let mut writer = FlushFailWriter::default();
-
-        let err = RolloutRecorder::persist_items_and_flush(
-            &mut writer,
-            &[RolloutItem::CompactionAttempt(attempt)],
-        )
-        .await
-        .expect_err("flush failure should be returned to the caller");
-
-        assert!(err.to_string().contains("synthetic flush failure"));
-        assert!(
-            !writer.buffer.is_empty(),
-            "writer should receive bytes before flush fails"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_flush_writer_propagates_flush_error() {
-        let mut writer = FlushFailWriter::default();
-
-        let err = RolloutRecorder::flush_writer(&mut writer)
-            .await
-            .expect_err("flush failure should be returned to the caller");
-
-        assert!(err.to_string().contains("synthetic flush failure"));
-    }
-
-    #[tokio::test]
-    async fn test_load_history_empty_file() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("empty.jsonl");
-
-        fs::write(&file_path, "").await.unwrap();
-
-        let items = RolloutRecorder::load_history(&file_path).await.unwrap();
-        assert!(items.is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_load_history_with_empty_lines() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("with_empty_lines.jsonl");
-
-        let content = r#"
-{"type":"message","role":"user","content":"Hello","tool_name":null,"timestamp":"2026-01-29T14:30:55Z"}
-
-{"type":"message","role":"assistant","content":"Hi!","tool_name":null,"timestamp":"2026-01-29T14:30:56Z"}
-"#;
-
-        fs::write(&file_path, content).await.unwrap();
-
-        let items = RolloutRecorder::load_history(&file_path).await.unwrap();
-        assert_eq!(items.len(), 2);
-    }
-
-    #[tokio::test]
-    async fn test_load_history_with_invalid_lines() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("with_invalid.jsonl");
-
-        let content = r#"{"type":"message","role":"user","content":"Valid","tool_name":null,"timestamp":"2026-01-29T14:30:55Z"}
-this is not valid json
-{"type":"message","role":"assistant","content":"Also valid","tool_name":null,"timestamp":"2026-01-29T14:30:56Z"}
-"#;
-
-        fs::write(&file_path, content).await.unwrap();
-
-        let error = RolloutRecorder::load_history(&file_path).await.unwrap_err();
-        assert!(error.to_string().contains("invalid current rollout record"));
-    }
-
-    #[tokio::test]
-    async fn test_load_history_ignores_torn_trailing_json_record() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("with_torn_tail.jsonl");
-        let content = concat!(
-            "{\"type\":\"message\",\"role\":\"user\",\"content\":\"Valid\",",
-            "\"tool_name\":null,\"timestamp\":\"2026-01-29T14:30:55Z\"}\n",
-            "{\"type\":\"message\",\"role\":\"assistant\",\"content\":"
-        );
-
-        fs::write(&file_path, content).await.unwrap();
-
-        let items = RolloutRecorder::load_history(&file_path).await.unwrap();
-        assert_eq!(items.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_load_history_ignores_torn_trailing_utf8_record() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("with_torn_utf8_tail.jsonl");
-        let mut content = concat!(
-            "{\"type\":\"message\",\"role\":\"user\",\"content\":\"Valid\",",
-            "\"tool_name\":null,\"timestamp\":\"2026-01-29T14:30:55Z\"}\n",
-            "{\"type\":\"message\",\"role\":\"assistant\",\"content\":\""
-        )
-        .as_bytes()
-        .to_vec();
-        content.extend_from_slice(&[0xe2, 0x82]);
-
-        fs::write(&file_path, content).await.unwrap();
-
-        let items = RolloutRecorder::load_history(&file_path).await.unwrap();
-        assert_eq!(items.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_load_history_rejects_non_torn_invalid_trailing_record() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("with_invalid_tail.jsonl");
-        let content = concat!(
-            "{\"type\":\"message\",\"role\":\"user\",\"content\":\"Valid\",",
-            "\"tool_name\":null,\"timestamp\":\"2026-01-29T14:30:55Z\"}\n",
-            "not json"
-        );
-
-        fs::write(&file_path, content).await.unwrap();
-
-        let error = RolloutRecorder::load_history(&file_path).await.unwrap_err();
-        assert!(error.to_string().contains("invalid current rollout record"));
-    }
-
-    #[tokio::test]
-    async fn test_load_history_accepts_valid_final_record_without_newline() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("valid_without_newline.jsonl");
-        let content = concat!(
-            "{\"type\":\"message\",\"role\":\"user\",\"content\":\"First\",",
-            "\"tool_name\":null,\"timestamp\":\"2026-01-29T14:30:55Z\"}\n",
-            "{\"type\":\"message\",\"role\":\"assistant\",\"content\":\"Second\",",
-            "\"tool_name\":null,\"timestamp\":\"2026-01-29T14:30:56Z\"}"
-        );
-
-        fs::write(&file_path, content).await.unwrap();
-
-        let items = RolloutRecorder::load_history(&file_path).await.unwrap();
-        assert_eq!(items.len(), 2);
-    }
-
-    #[tokio::test]
-    async fn test_load_history_file_not_found() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("nonexistent.jsonl");
-
-        let result = RolloutRecorder::load_history(&file_path).await;
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_rollout_item_serialization() {
-        let meta = RolloutItem::AgentMachineMeta(AgentMachineMeta {
-            rollout_id: "rollout-123".to_string(),
-            process_path: "/proc/7".to_string(),
-            started_at: "2026-01-29T14:30:52Z".to_string(),
-            cwd: "/test".to_string(),
-            model: "gemini-test".to_string(),
-            reasoning_effort: None,
-        });
-
-        let json = serde_json::to_string(&meta).unwrap();
-        assert!(json.contains("agent_machine_meta"));
-        assert!(json.contains("rollout-123"));
-        assert!(json.contains("/proc/7"));
-        assert!(json.contains("gemini-test"));
-
-        let deserialized: RolloutItem = serde_json::from_str(&json).unwrap();
-        match deserialized {
-            RolloutItem::AgentMachineMeta(m) => assert_eq!(m.rollout_id, "rollout-123"),
-            _ => panic!("Expected AgentMachineMeta"),
-        }
-    }
-
-    #[test]
-    fn test_message_record_serialization() {
-        let msg = MessageRecord {
-            role: "user".to_string(),
-            content: Some("Hello".to_string()),
-            tool_name: None,
-            message: None,
-            timestamp: "2026-01-29T14:30:55Z".to_string(),
-        };
-
-        let json = serde_json::to_string(&msg).unwrap();
-        assert!(json.contains("user"));
-        assert!(json.contains("Hello"));
-
-        let deserialized: MessageRecord = serde_json::from_str(&json).unwrap();
-        assert_eq!(deserialized.role, "user");
-        assert_eq!(deserialized.content, Some("Hello".to_string()));
-        assert!(deserialized.message.is_none());
-    }
-
-    #[test]
-    fn test_turn_context_item_serialization() {
-        let ctx = TurnContextItem {
-            model: "gemini-2.0-flash".to_string(),
-            reasoning_effort: None,
-            system_prompt: "System".to_string(),
-            context_items: vec![ContextItemRecord {
-                id: "onboarding".to_string(),
-                kind: "static".to_string(),
-                title: "Onboarding".to_string(),
-                content: "Steps".to_string(),
-                fingerprint: "abcd1234".to_string(),
-            }],
-            tools: vec!["web_search".to_string()],
-            memory_enabled: true,
-            active_skills: vec!["skill-1".to_string()],
-            reference_context: Some(ReferenceContextSnapshotRecord {
-                revision: 3,
-                changed: true,
-                reordered: false,
-                added: 1,
-                updated: 0,
-                removed: 0,
-            }),
-            timestamp: "2026-01-29T14:30:56Z".to_string(),
-        };
-
-        let json = serde_json::to_string(&ctx).unwrap();
-        assert!(json.contains("gemini-2.0-flash"));
-        assert!(json.contains("onboarding"));
-
-        let deserialized: TurnContextItem = serde_json::from_str(&json).unwrap();
-        assert_eq!(deserialized.model, "gemini-2.0-flash");
-        assert_eq!(deserialized.context_items[0].id, "onboarding");
-        assert_eq!(
-            deserialized.reference_context.as_ref().map(|r| r.revision),
-            Some(3)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_record_turn_context_persists_reasoning_effort() {
-        let temp_dir = TempDir::new().unwrap();
-        let recorder =
-            RolloutRecorder::new_in_dir("machine-123", "gemini-2.0-flash", temp_dir.path())
-                .await
-                .unwrap();
-
-        recorder
-            .record_turn_context(
-                "gemini-2.0-flash",
-                Some(alan_agent_protocol::ReasoningEffort::High),
-                "System",
-                vec![],
-                vec!["web_search".to_string()],
-                true,
-                vec![],
-                None,
-            )
-            .await
-            .unwrap();
-
-        let items = RolloutRecorder::load_history(recorder.path())
-            .await
-            .unwrap();
-        let persisted_effort = items.into_iter().find_map(|item| match item {
-            RolloutItem::TurnContext(ctx) => ctx.reasoning_effort,
-            _ => None,
-        });
-        assert_eq!(
-            persisted_effort,
-            Some(alan_agent_protocol::ReasoningEffort::High)
-        );
-    }
-
-    #[test]
-    fn test_turn_context_item_deserializes_without_reference_context_metadata() {
-        let json = r#"{
-            "model":"gemini-2.0-flash",
-            "system_prompt":"System",
-            "context_items":[],
-            "tools":["web_search"],
-            "memory_enabled":true,
-            "active_skills":[],
-            "timestamp":"2026-01-29T14:30:56Z"
-        }"#;
-
-        let deserialized: TurnContextItem = serde_json::from_str(json).unwrap();
-        assert!(deserialized.reference_context.is_none());
-    }
-
-    #[test]
-    fn test_compacted_item_serialization() {
-        let item = CompactedItem {
-            message: "Summary".to_string(),
-            attempt_id: Some("attempt-123".to_string()),
-            trigger: Some(CompactionTrigger::Manual),
-            reason: Some(CompactionReason::ExplicitRequest),
-            focus: Some("preserve todos".to_string()),
-            input_messages: Some(24),
-            output_messages: Some(8),
-            input_tokens: Some(1200),
-            output_tokens: Some(400),
-            duration_ms: Some(35),
-            retry_count: Some(1),
-            result: Some(CompactionResult::Retry),
-            reference_context_revision: Some(3),
-            timestamp: "2026-01-29T14:31:00Z".to_string(),
-        };
-
-        let json = serde_json::to_string(&item).unwrap();
-        assert!(json.contains("Summary"));
-        assert!(json.contains("\"manual\""));
-        assert!(json.contains("\"explicit_request\""));
-        assert!(json.contains("\"preserve todos\""));
-
-        let deserialized: CompactedItem = serde_json::from_str(&json).unwrap();
-        assert_eq!(deserialized.message, "Summary");
-        assert_eq!(deserialized.attempt_id.as_deref(), Some("attempt-123"));
-        assert_eq!(deserialized.trigger, Some(CompactionTrigger::Manual));
-        assert_eq!(deserialized.reason, Some(CompactionReason::ExplicitRequest));
-        assert_eq!(deserialized.focus.as_deref(), Some("preserve todos"));
-        assert_eq!(deserialized.reference_context_revision, Some(3));
-    }
-
-    #[test]
-    fn test_compaction_attempt_item_serialization() {
-        let attempt = CompactionAttemptSnapshot {
-            attempt_id: "attempt-123".to_string(),
-            submission_id: Some("sub-456".to_string()),
-            request: alan_agent_protocol::CompactionRequestMetadata {
-                mode: alan_agent_protocol::CompactionMode::Manual,
-                trigger: CompactionTrigger::Manual,
-                reason: CompactionReason::ExplicitRequest,
-                focus: Some("preserve todos".to_string()),
-            },
-            result: CompactionResult::Retry,
-            pressure_level: None,
-            memory_flush_attempt_id: None,
-            input_messages: Some(12),
-            output_messages: Some(4),
-            input_prompt_tokens: Some(900),
-            output_prompt_tokens: Some(300),
-            retry_count: 1,
-            tape_mutated: true,
-            warning_message: None,
-            error_message: None,
-            failure_streak: None,
-            reference_context_revision_before: Some(3),
-            reference_context_revision_after: Some(3),
-            timestamp: "2026-01-29T14:31:00Z".to_string(),
-        };
-
-        let json = serde_json::to_string(&RolloutItem::CompactionAttempt(attempt)).unwrap();
-        assert!(json.contains("\"compaction_attempt\""));
-        assert!(json.contains("\"attempt-123\""));
-        assert!(json.contains("\"retry\""));
-
-        let deserialized: RolloutItem = serde_json::from_str(&json).unwrap();
-        match deserialized {
-            RolloutItem::CompactionAttempt(snapshot) => {
-                assert_eq!(snapshot.attempt_id, "attempt-123");
-                assert_eq!(snapshot.submission_id.as_deref(), Some("sub-456"));
-                assert_eq!(snapshot.result, CompactionResult::Retry);
-                assert_eq!(snapshot.retry_count, 1);
-            }
-            other => panic!("expected compaction attempt item, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn test_tool_call_record_serialization() {
-        let tool = ToolCallRecord {
-            name: "web_search".to_string(),
-            arguments: serde_json::json!({"query": "test"}),
-            result: serde_json::json!({"found": 5}),
-            result_digest: Some("digest-1".to_string()),
-            result_preview: Some("found".to_string()),
-            redaction: Some(ToolPayloadRedactionSummary {
-                redacted_fields: 1,
-                truncated_values: 0,
-            }),
-            success: true,
-            audit: None,
-            timestamp: "2026-01-29T14:31:02Z".to_string(),
-        };
-
-        let json = serde_json::to_string(&tool).unwrap();
-        assert!(json.contains("web_search"));
-        assert!(json.contains("true"));
-
-        let deserialized: ToolCallRecord = serde_json::from_str(&json).unwrap();
-        assert_eq!(deserialized.name, "web_search");
-        assert!(deserialized.success);
-        assert_eq!(deserialized.result_digest.as_deref(), Some("digest-1"));
-        assert_eq!(deserialized.result_preview.as_deref(), Some("found"));
-        assert_eq!(
-            deserialized.redaction,
-            Some(ToolPayloadRedactionSummary {
-                redacted_fields: 1,
-                truncated_values: 0,
-            })
-        );
-    }
-
-    #[test]
-    fn test_build_durable_tool_payload_redacts_sensitive_headers() {
-        let durable = build_durable_tool_payload(&serde_json::json!({
-            "github_token": "github-secret",
-            "session_token": "machine-secret",
-            "status": 200,
-            "headers": {
-                "set-cookie": "machine=secret",
-                "authorization": "Bearer top-secret",
-                "content-type": "application/json"
-            }
-        }));
-
-        assert_eq!(
-            durable.payload["github_token"],
-            serde_json::json!("[REDACTED reason=secret_key]")
-        );
-        assert_eq!(
-            durable.payload["session_token"],
-            serde_json::json!("[REDACTED reason=secret_key]")
-        );
-        assert_eq!(
-            durable.payload["headers"]["set-cookie"],
-            serde_json::json!("[REDACTED reason=secret_key]")
-        );
-        assert_eq!(
-            durable.payload["headers"]["authorization"],
-            serde_json::json!("[REDACTED reason=secret_key]")
-        );
-        assert_eq!(
-            durable.payload["headers"]["content-type"],
-            serde_json::json!("application/json")
-        );
-        assert_eq!(
-            durable.redaction,
-            Some(ToolPayloadRedactionSummary {
-                redacted_fields: 4,
-                truncated_values: 0,
-            })
-        );
-    }
-
-    #[test]
-    fn test_build_durable_tool_payload_redacts_secrets_embedded_in_strings() {
-        let durable = build_durable_tool_payload(&serde_json::json!({
-            "output": "api_key=embedded-secret\nAuthorization: Bearer embedded-token"
-        }));
-
-        let output = durable.payload["output"]
-            .as_str()
-            .expect("output should remain a string");
-        assert!(!output.contains("embedded-secret"));
-        assert!(!output.contains("embedded-token"));
-        assert!(output.contains("[REDACTED reason=secret_key]"));
-        assert!(!durable.digest.is_empty());
-        assert!(
-            durable
-                .preview
-                .as_deref()
-                .is_none_or(|preview| !preview.contains("embedded-secret"))
-        );
-    }
-
-    #[test]
-    fn test_build_durable_tool_payload_truncates_large_strings() {
-        let durable = build_durable_tool_payload(&serde_json::json!({
-            "body": "x".repeat(2000)
-        }));
-
-        let body = durable.payload["body"]
-            .as_str()
-            .expect("body should stay a string");
-        assert!(body.contains("...[truncated]"));
-        assert_eq!(
-            durable.redaction,
-            Some(ToolPayloadRedactionSummary {
-                redacted_fields: 0,
-                truncated_values: 1,
-            })
-        );
-    }
-
-    #[test]
-    fn test_build_durable_tool_payload_preserves_bounded_evidence_projection_preview() {
-        let projection = crate::evidence::project_evidence_payload(
-            &serde_json::json!({
-                "output": "x".repeat(crate::evidence::MAX_INLINE_EVIDENCE_BYTES + 1)
-            }),
-            None,
-            Vec::new(),
-            Some("reference_unresolvable".to_string()),
-        );
-        let original_preview = projection["preview"].as_str().unwrap();
-        assert!(original_preview.len() > DURABLE_PAYLOAD_MAX_STRING_CHARS);
-
-        let durable = build_durable_tool_payload(&projection);
-
-        assert_eq!(durable.payload["preview"], projection["preview"]);
-        assert_eq!(
-            durable.payload["truncation"]["preview_bytes"],
-            serde_json::json!(original_preview.len())
-        );
-        assert!(durable.redaction.is_none());
-    }
-
-    #[test]
-    fn test_effect_record_serialization() {
-        let effect = EffectRecord {
-            effect_id: "ef-1".to_string(),
-            process_path: "/proc/7".to_string(),
-            tool_call_id: "call-1".to_string(),
-            idempotency_key: "idem-1".to_string(),
-            effect_type: "file".to_string(),
-            request_fingerprint: "fp-1".to_string(),
-            result_digest: Some("digest-1".to_string()),
-            result_payload: Some(serde_json::json!({"ok": true})),
-            status: EffectStatus::Applied,
-            applied_at: Some("2026-03-03T10:00:00Z".to_string()),
-            reason: None,
-            dedupe_hit: false,
-            timestamp: "2026-03-03T10:00:01Z".to_string(),
-        };
-
-        let json = serde_json::to_string(&effect).unwrap();
-        assert!(json.contains("ef-1"));
-        assert!(json.contains("applied"));
-
-        let deserialized: EffectRecord = serde_json::from_str(&json).unwrap();
-        assert_eq!(deserialized.effect_id, "ef-1");
-        assert_eq!(deserialized.status, EffectStatus::Applied);
-    }
-
-    #[tokio::test]
-    async fn test_record_tool_call_redacts_sensitive_values_before_persisting() {
-        let temp_dir = TempDir::new().unwrap();
-        let recorder =
-            RolloutRecorder::new_in_dir("test-tool-redaction", "gemini-2.0-flash", temp_dir.path())
-                .await
-                .unwrap();
-
-        recorder
-            .record_tool_call_with_audit(
-                "web_fetch",
-                serde_json::json!({
-                    "headers": {
-                        "authorization": "Bearer top-secret"
-                    }
-                }),
-                serde_json::json!({
-                    "headers": {
-                        "set-cookie": "machine=super-secret"
-                    }
-                }),
-                true,
-                None,
-            )
-            .await
-            .unwrap();
-
-        let items = RolloutRecorder::load_history(recorder.path())
-            .await
-            .unwrap();
-        let tool_call = items.into_iter().find_map(|item| match item {
-            RolloutItem::ToolCall(record) => Some(record),
-            _ => None,
-        });
-
-        let record = tool_call.expect("tool call should be persisted");
-        assert_eq!(
-            record.arguments["headers"]["authorization"],
-            serde_json::json!("[REDACTED reason=secret_key]")
-        );
-        assert_eq!(
-            record.result["headers"]["set-cookie"],
-            serde_json::json!("[REDACTED reason=secret_key]")
-        );
-        assert!(
-            record.result_digest.is_some(),
-            "durable tool call should persist a result digest"
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_record_serialization() {
-        let cp = CheckpointRecord {
-            checkpoint_id: "cp-123".to_string(),
-            checkpoint_type: "supplier_list".to_string(),
-            summary: "Found 5 suppliers".to_string(),
-            choice: Some("approved".to_string()),
-            knowledge_root: Some("sha256:abc123".to_string()),
-            timestamp: "2026-01-29T14:35:00Z".to_string(),
-        };
-
-        let json = serde_json::to_string(&cp).unwrap();
-        assert!(json.contains("cp-123"));
-        assert!(json.contains("approved"));
-        assert!(json.contains("sha256:abc123"));
-
-        let deserialized: CheckpointRecord = serde_json::from_str(&json).unwrap();
-        assert_eq!(deserialized.checkpoint_id, "cp-123");
-        assert_eq!(deserialized.choice, Some("approved".to_string()));
-        assert_eq!(
-            deserialized.knowledge_root.as_deref(),
-            Some("sha256:abc123")
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_record_without_choice() {
-        let cp = CheckpointRecord {
-            checkpoint_id: "cp-456".to_string(),
-            checkpoint_type: "requirements".to_string(),
-            summary: "Requirements gathered".to_string(),
-            choice: None,
-            knowledge_root: None,
-            timestamp: "2026-01-29T14:36:00Z".to_string(),
-        };
-
-        let json = serde_json::to_string(&cp).unwrap();
-        assert!(json.contains("cp-456"));
-        assert!(json.contains("null"));
-        assert!(
-            !json.contains("knowledge_root"),
-            "missing root stays omitted for legacy-compatible checkpoints: {json}"
-        );
-
-        let deserialized: CheckpointRecord = serde_json::from_str(&json).unwrap();
-        assert!(deserialized.choice.is_none());
-        assert!(deserialized.knowledge_root.is_none());
-    }
-
-    #[test]
-    fn test_event_record_serialization() {
-        let event = EventRecord {
-            event_type: "thinking".to_string(),
-            payload: serde_json::json!({"message": "Analyzing..."}),
-            timestamp: "2026-01-29T14:37:00Z".to_string(),
-        };
-
-        let json = serde_json::to_string(&event).unwrap();
-        assert!(json.contains("thinking"));
-        assert!(json.contains("Analyzing..."));
-
-        let deserialized: EventRecord = serde_json::from_str(&json).unwrap();
-        assert_eq!(deserialized.event_type, "thinking");
-    }
-
-    #[test]
-    fn test_rollout_recorder_clone() {
-        // Just test that Clone is implemented correctly
-        let _temp_dir = TempDir::new().unwrap();
-        let _path = _temp_dir.path().join("test.jsonl");
-
-        // Create a minimal recorder for testing clone
-        // Note: We can't easily create a full recorder without async, but we can verify the types
-        fn check_clone<T: Clone>(_: T) {}
-
-        // Check that RolloutItem implements Clone
-        let item = RolloutItem::Message(MessageRecord {
-            role: "user".to_string(),
-            content: Some("test".to_string()),
-            tool_name: None,
-            message: None,
-            timestamp: "2026-01-29T14:30:55Z".to_string(),
-        });
-        check_clone(item);
-    }
-
-    #[tokio::test]
-    async fn test_load_history_checkpoint_item() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("checkpoint.jsonl");
-
-        let content = r#"{"type":"checkpoint","checkpoint_id":"cp-test","checkpoint_type":"supplier_list","summary":"Test summary","choice":"approved","timestamp":"2026-01-29T14:35:00Z"}"#;
-
-        fs::write(&file_path, content).await.unwrap();
-
-        let items = RolloutRecorder::load_history(&file_path).await.unwrap();
-        assert_eq!(items.len(), 1);
-
-        match &items[0] {
-            RolloutItem::Checkpoint(cp) => {
-                assert_eq!(cp.checkpoint_id, "cp-test");
-                assert_eq!(cp.checkpoint_type, "supplier_list");
-                assert_eq!(cp.summary, "Test summary");
-                assert_eq!(cp.choice, Some("approved".to_string()));
-            }
-            _ => panic!("Expected Checkpoint"),
-        }
-    }
-
-    #[tokio::test]
-    async fn test_load_history_event_item() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("event.jsonl");
-
-        let content = r#"{"type":"event","event_type":"thinking","payload":{"message":"Thinking..."},"timestamp":"2026-01-29T14:37:00Z"}"#;
-
-        fs::write(&file_path, content).await.unwrap();
-
-        let items = RolloutRecorder::load_history(&file_path).await.unwrap();
-        assert_eq!(items.len(), 1);
-
-        match &items[0] {
-            RolloutItem::Event(evt) => {
-                assert_eq!(evt.event_type, "thinking");
-            }
-            _ => panic!("Expected Event"),
-        }
-    }
-}
+#[path = "rollout_tests.rs"]
+mod tests;
