+++ crates/agent-engine/src/rollout_tests.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+#[path = "rollout_recorder_tests.rs"]
+mod recorder;
+#[path = "rollout_record_tests.rs"]
+mod records;
