+++ scripts/rust-source-size-baseline.txt
@@ -2,7 +2,7 @@
 # A file may only stay equal to this maximum or shrink with this entry tightened.
 crates/agent-engine/skills/swebench/tooling/src/swebench_lite.rs 1389
 crates/agent-engine/src/config.rs 2556
-crates/agent-engine/src/rollout.rs 2266
+crates/agent-engine/src/rollout.rs 1200
 crates/agent-engine/src/skills/injector.rs 2490
 crates/agent-engine/src/skills/types.rs 3013
 crates/agent-engine/src/tools/reified_namespace.rs 2756
